+++ packages/common-resources/static-resources/src/i18n/en_US/common.ts
@@ -793,7 +793,7 @@ export default {
   labelLayerSwapUnderstandDes:
     "LayerSwap is a 3rd party App service provider to help move tokens from exchange to Loopring L2 directly. If you have any concerns regarding their service, please check out their <1>TOS</1>.",
   labelInvestAmmTitle: "AMM Pools",
-  labelInvestBalanceTitle: "My Investment",
+  labelInvestBalanceTitle: "My Investments",
   labelTransactionsLink: "Transactions",
   labelAMMTransactionsLink: "View Pool Transactions",
   labelNFTMintWrongCIDBtn: "Wrong MetaData format",
@@ -826,7 +826,7 @@ export default {
   labelForceWithdrawNoToken:
     "No token is detected from this address to operate",
   labelForceWithdrawBtn: "Force Withdraw",
-  labelInvestDefiTitle: "Defi Investment",
+  labelInvestDefiTitle: "ETH Staking",
   labelInvestDefDeposit: "Subscribe",
   labelInvestDefWithdraw: "Redeem",
   labelNFTDepositLabel: "Receive NFT",
@@ -845,13 +845,13 @@ export default {
   labelSecurity: "Security",
   labelFeeTitleList: "Fee",
   labelInvestOverviewTitle: "Overview",
-  labelTitleOverviewToken: "Investment Tokens",
+  labelTitleOverviewToken: "Total Investment Tokens",
   labelInvestType_AMM: "AMM Pool",
-  labelInvestType_DEFI: "Lido Stake",
+  labelInvestType_DEFI: "ETH Staking",
   labelInvestFlexible: "Flexible",
   labelInvestDuration: "Duration",
-  labelDefiOrderTable: "Defi History",
-  labelTitleMyInvestAvaiable: "My Tokens",
+  labelDefiOrderTable: "ETH Staking",
+  labelTitleMyInvestAvaiable: "My holding tokens",
   labelViewMore: "View more",
   labelInvestSuccess: "Successfully {{type}} {{symbol}}",
   labelInvestFailed: "Subscribe Failed",
