+++ packages/common-resources/static-resources/src/i18n/zh_CN/common.ts
@@ -814,7 +814,7 @@ export default {
   labelForceWithdrawNoToken:
     "No token is detected from this address to operate",
   labelForceWithdrawBtn: "Force Withdraw",
-  labelInvestDefiTitle: "Defi Invest",
+  labelInvestDefiTitle: "ETH Staking",
   labelInvestDefDeposit: "Invest",
   labelInvestDefWithdraw: "Redeem",
   labelNFTDepositLabel: "Receive NFT",
@@ -833,12 +833,12 @@ export default {
   labelSecurity: "Security",
   labelFeeTitleList: "Fee",
   labelInvestOverviewTitle: "Overview",
-  labelTitleOverviewToken: "Investment Tokens",
+  labelTitleOverviewToken: "Total Investment Tokens",
   labelInvestType_AMM: "AMM Pool",
-  labelInvestType_DEFI: "Lido Stake",
+  labelInvestType_DEFI: "ETH Staking",
   labelInvestFlexible: "Flexible",
   labelInvestDuration: "Duration",
   labelDefiOrderTable: "Defi History",
-  labelTitleMyInvestAvaiable: "My Tokens",
+  labelTitleMyInvestAvaiable: "Holding tokens",
   labelViewMore: "View more",
 };
